행 삭제 처리 api 에서 결과 반환

+++ src/main/kotlin/com/raillylinker/springboot_mvc_template/controllers/c8_service1_tk_v1_redisTest/C8Service1TkV1RedisTestService.kt
@@ -93,6 +93,14 @@ class C8Service1TkV1RedisTestService(
     ////
     @CustomRedisTransactional([Redis1_Test.TRANSACTION_NAME])
     fun api4(httpServletResponse: HttpServletResponse, key: String) {
+        val keyValue = redis1TestRepository.findKeyValue(key)
+
+        if (keyValue == null) {
+            httpServletResponse.status = HttpStatus.NO_CONTENT.value()
+            httpServletResponse.setHeader("api-result-code", "1")
+            return
+        }
+
         redis1TestRepository.deleteKeyValue(key)
 
         httpServletResponse.status = HttpStatus.OK.value()
